Animations were added

+++ Trigonometry/Trigonometry/Data/BiorythmModel.swift
@@ -11,27 +11,27 @@ import SwiftUI
 struct Biorhythm {
     let type: BiorhythmType
     let distance: Int = 33 // According to the HoR-24 33 days is a fixed value
-    var firstDay: Int {
+    var firstDay: Double {
         didSet {
-            currentState = getStateForDay(dayFromBirth: Double(firstDay + 3))
+            currentState = getStateForDay(dayFromBirth: firstDay + 3)
         }
     }
     var currentState = 0.0
     let name: String
     
-    init(type: BiorhythmType, name: String, firstDay: Int = DateHelper().daysFromBirthDay - 3) {
+    init(type: BiorhythmType, name: String, firstDay: Double = DateHelper().daysFromBirthDay - 3) {
         self.type = type
         self.name = name
         self.firstDay = firstDay
-        currentState = getStateForDay(dayFromBirth: Double(firstDay + 3))
+        currentState = getStateForDay(dayFromBirth: firstDay + 3)
     }
     
     func getStateForDay(dayFromBirth: Double) -> Double {
         type == .overall
-        ? sin(2 * Double.pi * Double(dayFromBirth) / BiorhythmType.intellectual.rawValue) +
-        sin(2 * Double.pi * Double(dayFromBirth) / BiorhythmType.physical.rawValue) +
-        sin(2 * Double.pi * Double(dayFromBirth) / BiorhythmType.emotional.rawValue)
-        : sin(2 * Double.pi * Double((dayFromBirth)) / type.rawValue)
+        ? sin(2 * Double.pi * dayFromBirth / BiorhythmType.intellectual.rawValue) +
+        sin(2 * Double.pi * dayFromBirth / BiorhythmType.physical.rawValue) +
+        sin(2 * Double.pi * dayFromBirth / BiorhythmType.emotional.rawValue)
+        : sin(2 * Double.pi * dayFromBirth / type.rawValue)
     }
     
     func getStateInPercent() -> Int {
