+++ Trigonometry/Trigonometry/Data/DateHelper.swift
@@ -8,15 +8,15 @@
 import Foundation
 
 extension Date {
-    func distanceFromTodayInDays(dayBefore: Date) -> Int {
-        Calendar.current.dateComponents([.day], from: dayBefore, to: Date()).day ?? 0
+    func distanceFromTodayInDays(dayBefore: Date) -> Double {
+        Double(Calendar.current.dateComponents([.day], from: dayBefore, to: Date()).day ?? 0)
     }
 }
 
 final class DateHelper: ObservableObject {
     let today = Date()
     let bday = Calendar.current.date(from: DateComponents(year: 2017, month: 9, day: 17))! // 2DO: replace with Environment's date
-    var daysFromBirthDay: Int { today.distanceFromTodayInDays(dayBefore: bday) }
+    var daysFromBirthDay: Double { today.distanceFromTodayInDays(dayBefore: bday) }
     
     static func getAxisNamesForDay(for day: Date) -> String {
         let formatter = DateFormatter()
